started exercise05
I am going to do a random but controlled screen resizing, which happens when the user presses any key. At the beginning, it says "press any key and try to find the logic", so the user will press any key and try to find the logic in relation to the key pressed and the screen resizing. Then, a short while after, a message will appear saying something like "Tricked you!! There is no logic :)"