Worked in class with Pippin
A lot of changes : we separated the toys (they were all on top of each other) by regenerating a random value for each. We fixed the timer and the toys' behavior so that theyre only displayed when they fall, and they fall every 3 seconds or so. We also cleaned my code a bit, moving codes around and deleting some things. We also created a code that allows us to know which elf is dropping it.

NOW I HAVE TO : Make sure the toys fall randomly, by putting a timer in the toy class as well.